♻️ :: schoolId Int -> Long

+++ data/src/main/java/com/semicolon/data/remote/response/school/SearchSchoolResponse.kt
@@ -7,7 +7,7 @@ data class SearchSchoolResponse(
     @SerializedName("search_school_list") val schoolList: List<SchoolInfo>
 ) {
     data class SchoolInfo(
-        @SerializedName("school_id") val agencyCode: Int,
+        @SerializedName("school_id") val agencyCode: Long,
         @SerializedName("school_name") val schoolName: String,
         @SerializedName("logo_image_url") val logoImageUrl: String,
     )
